feat(widget): #MA-526 register side-widget state in session
